requirements.txt
celery==5.1.2
django==3.2.4
django-filter==2.4.0
django-rest-framework==0.1.0
gunicorn==20.1.0
mock==4.0.3
pandas==1.3.0
redis==3.5.3
